% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tidymass.R
\name{install_tidymass}
\alias{install_tidymass}
\title{install_tidymass}
\usage{
install_tidymass(
  from = c("github", "gitee"),
  force = FALSE,
  upgrade = "never",
  dependencies = NA,
  demo_data = TRUE,
  which_package = c("all", "metID", "lipidflow", "demoData", "metPath"),
  ...
)
}
\arguments{
\item{from}{From github or gitee, if you are in China, try to set this as "gitee".}

\item{force}{Force installation, even if the remote state has not changed since the previous install.}

\item{upgrade}{One of "default", "ask", "always", or "never".
"default" respects the value of the R_REMOTES_UPGRADE environment variable if set,
and falls back to "ask" if unset. "ask" prompts the user for which out of date
packages to upgrade. For non-interactive sessions "ask" is equivalent to "always".
TRUE and FALSE are also accepted and correspond to "always" and "never" respectively.}

\item{dependencies}{Which dependencies do you want to check? Can be a character vector
(selecting from "Depends", "Imports", "LinkingTo", "Suggests", or "Enhances"),
or a logical vector.TRUE is shorthand for "Depends", "Imports", "LinkingTo"
and "Suggests". NA is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. FALSE is shorthand for no dependencies
(i.e. just check this package, not its dependencies).
The value "soft" means the same as TRUE, "hard" means the same as NA.}

\item{demo_data}{Install demo_data package or not.}

\item{which_package}{What packages you want to install? Default is all. You can set it as a character vector.}

\item{...}{Other parameters from devtools::install_github() or devtools::install_git()}
}
\description{
Install all packages in tidymass.
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
