% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-launch_shiny.R
\name{launch_metid_shiny}
\alias{launch_metid_shiny}
\title{Launch metid Shiny Application}
\usage{
launch_metid_shiny()
}
\value{
None.
}
\description{
This function launches the metid Shiny application included in the 'metid' package.
It locates the Shiny app directory within the installed package and runs the app.
}
\note{
This function does not take any arguments.
Ensure that the 'metid' package is correctly installed with the Shiny app included.
}
