% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-launch_tidymass_shiny.R
\name{launch_tidymass_shiny}
\alias{launch_tidymass_shiny}
\title{Launch tidymass Shiny Application}
\usage{
launch_tidymass_shiny()
}
\value{
None.
}
\description{
This function launches the tidymass Shiny application included in the 'tidymass' package.
It locates the Shiny app directory within the installed package and runs the app.
}
\note{
This function does not take any arguments.
Ensure that the 'tidymass' package is correctly installed with the Shiny app included.
}
