% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass.R
\name{tidymass}
\alias{tidymass}
\title{Show the base information of tidymass pacakge}
\usage{
tidymass()
}
\value{
A ASCII log of tidymass
}
\description{
Show the base information of tidymass pacakge.
\lifecycle{maturing}
}
\examples{
tidymass()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
