% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_update.R
\name{tidymass_deps}
\alias{tidymass_deps}
\title{List all tidymass dependencies}
\usage{
tidymass_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidymass packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all tidymass dependencies
}
