% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-tidymass_logo.R
\name{tidymass_logo}
\alias{tidymass_logo}
\title{Show the tidymass logo}
\usage{
tidymass_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Defaults to \code{TRUE} on
UTF-8 platforms.}
}
\value{
A character vector containing the formatted tidymass logo.
}
\description{
Display the tidymass logo using ASCII or Unicode characters.
}
\examples{
tidymass_logo()
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
