% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidymass_packages}
\alias{tidymass_packages}
\title{List all packages in the tidymass}
\usage{
tidymass_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidymass in the list?}
}
\description{
List all packages in the tidymass
}
\examples{
tidymass_packages()
}
