% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_update.R
\name{tidymass_update}
\alias{tidymass_update}
\title{Update tidymass packages}
\usage{
tidymass_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidymass packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
This will check to see if all tidymass packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidymass_update()
}
}
