% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_logo.R
\name{tidymass_logo}
\alias{tidymass_logo}
\title{Show the logo tidymass.}
\usage{
tidymass_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\value{
A ASCII log of tidymass
}
\description{
The tidymass logo, using ASCII or Unicode characters
}
\examples{
tidymass_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
