% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_update.R
\name{update_tidymass}
\alias{update_tidymass}
\title{update_tidymass}
\usage{
update_tidymass(
  packages = c("core", "all"),
  from = c("github", "gitlab", "gitee")
)
}
\arguments{
\item{packages}{core or all packages in tidymass. "core" means all the core
packages in tidymass and "all" means all the packages in tidymass.}

\item{from}{github, gitlab or gitee.}
}
\description{
Update packages in tidymass.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
